changed iterator to uint8_t

+++ lib/PortSchema/examples/example_port_schema/Ports.cpp
@@ -1,7 +1,7 @@
 #include "Ports.h"
 
 portSchema getPort(uint8_t port_number) {
-    for (int i = 0; i < (sizeof(all_ports) / sizeof(portSchema)); i++) {
+    for (uint8_t i = 0; i < (sizeof(all_ports) / sizeof(portSchema)); i++) {
         if (all_ports[i].port_number == port_number) {
             return all_ports[i];
         }
